FEAT Admin and customer login parallel credintial checking added
